+++ src/components/UserDetailSections/UserAffiliations/UserAffiliations.test.js
@@ -1,16 +1,35 @@
-import { screen } from '@testing-library/react';
-
+import {
+  screen,
+  waitForElementToBeRemoved,
+} from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import {
+  QueryClient,
+  QueryClientProvider,
+} from 'react-query';
 import renderWithRouter from 'helpers/renderWithRouter';
 
 import affiliations from '../../../../test/jest/fixtures/affiliations';
-import { useUserAffiliations } from '../../../hooks';
+import {
+  useConsortiumTenants,
+  useUserAffiliations,
+  useUserAffiliationsMutation
+} from '../../../hooks';
 import UserAffiliations from './UserAffiliations';
+import { getResponseErrors } from './util';
+
+const queryClient = new QueryClient();
 
 jest.unmock('@folio/stripes/components');
 jest.mock('../../../hooks', () => ({
   ...jest.requireActual('../../../hooks'),
-  useUserAffiliations: jest.fn(() => ({ affiliations: [], totalRecords: 0 })),
-  useUserAffiliationsMutation: jest.fn(() => ({ handleAssignment: jest.fn(), isLoading: false })),
+  useConsortiumTenants: jest.fn(),
+  useUserAffiliations: jest.fn(),
+  useUserAffiliationsMutation: jest.fn(),
+}));
+
+jest.mock('./util', () => ({
+  getResponseErrors: jest.fn(() => []),
 }));
 
 const defaultProps = {
@@ -18,24 +37,33 @@ const defaultProps = {
   expanded: true,
   onToggle: jest.fn(),
   userId: 'userId',
+  userName: 'mobius',
 };
 
-const renderPatronBlock = (props = {}) => renderWithRouter(
-  <UserAffiliations
-    {...defaultProps}
-    {...props}
-  />
+const tenants = affiliations.map(({ tenantId, tenantName }) => ({ id: tenantId, name: tenantName }));
+
+const renderUserAffiliations = (props = {}) => renderWithRouter(
+  <QueryClientProvider client={queryClient}>
+    <UserAffiliations
+      {...defaultProps}
+      {...props}
+    />
+  </QueryClientProvider>
 );
 
-describe('render UserAffiliations accordion component', () => {
+describe('UserAffiliations', () => {
+  const handleAssignment = jest.fn(() => Promise.resolve([]));
+
   beforeEach(() => {
+    useConsortiumTenants.mockClear().mockReturnValue({ tenants });
+    useUserAffiliationsMutation.mockClear().mockReturnValue({ handleAssignment, isLoading: false });
     useUserAffiliations
       .mockClear()
-      .mockReturnValue({ affiliations, totalRecords: affiliations.length });
+      .mockReturnValue({ affiliations, totalRecords: affiliations.length, isLoading: false, handleAssignment: () => [{}], refetch: () => {} });
   });
 
   it('should render a list of user affiliations', () => {
-    renderPatronBlock();
+    renderUserAffiliations();
 
     affiliations.map(({ tenantName }) => {
       return expect(screen.getByText(tenantName)).toBeInTheDocument();
@@ -43,10 +71,45 @@ describe('render UserAffiliations accordion component', () => {
   });
 
   it('should render primary affiliation list item with \'primary\' class', () => {
-    renderPatronBlock();
+    renderUserAffiliations();
 
     const primaryTenantListItem = screen.getByText(affiliations.find(({ isPrimary }) => isPrimary).tenantName);
-
     expect(primaryTenantListItem).toHaveClass('primary');
+  });
+
+  it.each`
+    status
+    ${'error'}
+    ${'success'}
+  `('should show $status message on click saveAndClose button', async ({ status }) => {
+    let mockErrorData = [];
+    if (status === 'error') {
+      mockErrorData = [{
+        'message': 'User with id [0c50701e-45ff-4a2e-bff0-11bd5610378d] has primary affiliation with tenant [mobius]. Primary Affiliation cannot be deleted',
+        'type': '-1',
+        'code': 'HAS_PRIMARY_AFFILIATION_ERROR'
+      },
+      {
+        'message': 'Some error message',
+        'type': '-1',
+        'code': 'GENERIC_ERROR'
+      }];
+    }
+
+    getResponseErrors.mockClear().mockReturnValue(mockErrorData);
+    renderUserAffiliations();
+
+    const assignButton = screen.getByText('ui-users.affiliations.section.action.edit');
+    userEvent.click(assignButton);
+    const listOfAssignedTenants = await screen.findAllByRole('checkbox', {
+      name: 'ui-users.affiliations.manager.modal.aria.assign',
+      checked: true,
+    });
+    expect(listOfAssignedTenants).toHaveLength(affiliations.length);
+    const saveAndCloseButton = screen.getByText('ui-users.saveAndClose');
+    userEvent.click(saveAndCloseButton);
+    await waitForElementToBeRemoved(() => screen.queryByText('ui-users.affiliations.manager.modal.title'));
+    expect(handleAssignment).toHaveBeenCalled();
+    expect(screen.queryByText('ui-users.saveAndClose')).toBeNull();
   });
 });