+++ src/hooks/useUserAffiliationsMutation/useUserAffiliationsMutation.js
@@ -1,14 +1,15 @@
 import chunk from 'lodash/chunk';
 import { useCallback } from 'react';
 import { useMutation } from 'react-query';
-
 import { useOkapiKy } from '@folio/stripes/core';
-
+import { uniqBy } from 'lodash';
 import {
   CONSORTIA_API,
   CONSORTIA_USER_TENANTS_API,
+  OKAPI_TENANT_HEADER,
 } from '../../constants';
 import useConsortium from '../useConsortium';
+import { getResponseErrors } from '../../components/UserDetailSections/UserAffiliations/util';
 
 const CHUNK_SIZE = 5;
 
@@ -33,6 +34,16 @@ const useUserAffiliationsMutation = () => {
     isLoading: isConsortiumLoading,
   } = useConsortium();
 
+  const api = ky.extend({
+    hooks: {
+      beforeRequest: [
+        request => {
+          request.headers.set(OKAPI_TENANT_HEADER, consortium.centralTenant);
+        },
+      ],
+    },
+  });
+
   const {
     mutateAsync: assignAffiliation,
     isLoading: isAssigningLoading,
@@ -43,7 +54,7 @@ const useUserAffiliationsMutation = () => {
         userId,
       };
 
-      return ky.post(
+      return api.post(
         `${CONSORTIA_API}/${consortium.id}/${CONSORTIA_USER_TENANTS_API}`,
         { json },
       );
@@ -57,10 +68,10 @@ const useUserAffiliationsMutation = () => {
     mutationFn: ({ tenantId, userId }) => {
       const searchParams = {
         tenantId,
-        userId
+        userId,
       };
 
-      return ky.delete(
+      return api.delete(
         `${CONSORTIA_API}/${consortium.id}/${CONSORTIA_USER_TENANTS_API}`,
         { searchParams },
       );
@@ -68,10 +79,26 @@ const useUserAffiliationsMutation = () => {
   });
 
   const handleAssignment = useCallback(async ({ added, removed }) => {
-    return Promise.allSettled([
+    const batchResponses = await Promise.allSettled([
       batchRequest(added, assignAffiliation),
       batchRequest(removed, unassignAffiliation),
     ]);
+
+    const errors = await getResponseErrors(batchResponses);
+    const uniqueErrorMessages = uniqBy(errors, 'message');
+    if (uniqueErrorMessages.length) {
+      return {
+        errors: uniqueErrorMessages,
+        success: false,
+        responses: batchResponses,
+      };
+    }
+
+    return {
+      errors: [],
+      success: true,
+      responses: batchResponses,
+    };
   }, [assignAffiliation, unassignAffiliation]);
 
   const isLoading = isConsortiumLoading || isAssigningLoading || isUnassigningLoading;